// BudgetApp/BudgetApp/ClientApp/src/app/swagger-generated/model/transactionViewModel.ts
/**
 * BudgetApp
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransactionViewModel { 
    transactionId?: number;
    transactionName?: string;
    category?: string;
    transactionStatus?: string;
    postedDate?: Date;
    upcomingDate?: Date;
    amount?: number;
}